Add a simple build script to bootstrap a build
This script works as instructions to a new developer as well as just
making everything a bit easier to trigger from editors, etc.

Yeah, the farce of having a build script to invoke a build script to
generate and invoke a buildscript to invoke a build is not lost on me.
I guess that's just part of doing C; generations of tools getting
slightly better at each time...